Downgrade to Minecraft 1.2.5

+++ src/main/java/net/rush/packets/packet/BlockChangePacket.java
@@ -6,6 +6,6 @@ public interface BlockChangePacket extends Packet {
     int getX();
     byte getY();
     int getZ();
-    short getBlockType();
+    byte getBlockType();
     byte getBlockMetadata();
 }
