+++ src/main/java/net/rush/packets/packet/UpdateHealthPacket.java
@@ -3,7 +3,7 @@ package net.rush.packets.packet;
 import net.rush.packets.Packet;
 
 public interface UpdateHealthPacket extends Packet {
-	float getHealth();
+    short getHealth();
     short getFood();
     float getSaturation();
 }
